Allow extending the http pipeline (#2)
Adds to the http.Bind message a function that allows extending the http pipeline. This function receives the standard pipeline and returns a possibly modified one.

Our main use case is adding support for the HAProxy proxy protocol in situations where a front-end load balancer is not able to pass the common X-Forwarded-For headers.